Messages Table component created
